seperate controller for avatar, also fixes bug( fixes: #123)
