+++ www/.eslintrc.json
@@ -7,6 +7,7 @@
   },
   "env": { "es6": true },
   "rules": {
-    "require-jsdoc": "off"
+    "require-jsdoc": "off",
+    "no-unused-vars": "error"
   }
 }
