+++ www/src/services/auth.ts
@@ -1,14 +1,14 @@
 import Network from './network';
 
 export default class Auth {
-  public static async login<T>(username: string, password: string): T {
+  public static async login<T>(username: string, password: string): Promise<T> {
     return Network.request<T>('POST', '/session', {
       username: username,
       password: password,
     });
   };
 
-  public static async resume<T>(): T {
+  public static async resume<T>(): Promise<T> {
     return Network.request<T>('GET', '/session');
   };
 
@@ -16,7 +16,7 @@ export default class Auth {
     return Network.request('DELETE', '/session');
   };
 
-  public static async register<T>(user: T, password: string): T {
+  public static async register<T>(user: T, password: string): Promise<T> {
     return Network.request<T>('POST', '/user', user); // TODO: Handle password.
   };
 }
